+++ tests/PidQueryBuilderTest.php
@@ -14,10 +14,7 @@ final class PidQueryBuilderTest extends TestCase
 {
     public function testBuildsAlternativeMdocAndSdJwtQueries(): void
     {
-        $query = (new PidQueryBuilder())->build(
-            [Claim::BIRTH_DATE, Claim::RESIDENT_CITY],
-            'Identity check',
-        );
+        $query = (new PidQueryBuilder())->build([Claim::BIRTH_DATE, Claim::RESIDENT_CITY]);
 
         $this->assertSame(PidQueryBuilder::MDOC_ID, $query['credentials'][0]['id']);
         $this->assertSame('mso_mdoc', $query['credentials'][0]['format']);
@@ -37,7 +34,7 @@ final class PidQueryBuilderTest extends TestCase
 
     public function testCanRequestMdocOnly(): void
     {
-        $query = (new PidQueryBuilder())->build([Claim::GIVEN_NAME], 'Name', RequestOptions::FORMAT_MDOC);
+        $query = (new PidQueryBuilder())->build([Claim::GIVEN_NAME], RequestOptions::FORMAT_MDOC);
 
         $this->assertCount(1, $query['credentials']);
         $this->assertSame('mso_mdoc', $query['credentials'][0]['format']);
@@ -45,7 +42,7 @@ final class PidQueryBuilderTest extends TestCase
 
     public function testBothFallsBackToSdJwtForSdJwtOnlyAttribute(): void
     {
-        $query = (new PidQueryBuilder())->build([Claim::RESIDENT_HOUSE_NUMBER], 'Address');
+        $query = (new PidQueryBuilder())->build([Claim::RESIDENT_HOUSE_NUMBER]);
 
         $this->assertCount(1, $query['credentials']);
         $this->assertSame(PidQueryBuilder::SD_JWT_ID, $query['credentials'][0]['id']);
@@ -56,18 +53,18 @@ final class PidQueryBuilderTest extends TestCase
     public function testRejectsAttributeUnavailableInExplicitFormat(): void
     {
         $this->expectException(UnknownClaim::class);
-        (new PidQueryBuilder())->build([Claim::RESIDENT_HOUSE_NUMBER], 'Address', RequestOptions::FORMAT_MDOC);
+        (new PidQueryBuilder())->build([Claim::RESIDENT_HOUSE_NUMBER], RequestOptions::FORMAT_MDOC);
     }
 
     public function testRejectsUnknownClaims(): void
     {
         $this->expectException(UnknownClaim::class);
-        (new PidQueryBuilder())->build(['not_a_pid_claim'], 'Nope');
+        (new PidQueryBuilder())->build(['not_a_pid_claim']);
     }
 
     public function testRejectsDuplicateClaims(): void
     {
         $this->expectException(\InvalidArgumentException::class);
-        (new PidQueryBuilder())->build([Claim::GIVEN_NAME, Claim::GIVEN_NAME], 'Name');
+        (new PidQueryBuilder())->build([Claim::GIVEN_NAME, Claim::GIVEN_NAME]);
     }
 }